Export more components
The index file now exports a number of named components rather than just
the "root" component.  This will allow Flight Launch for instance to use
just the ReactTerminal component much more easily than before.
